# Optimize parsing speed by reducing reliance on exceptions