--- a/src/generate-types.ts
+++ b/src/generate-types.ts
@@ -634,55 +634,18 @@ export function run(options: Options) {
     );
   }
 
-  function generateClassConstructor(key: string) {
+  function generateClassConstructor() {
     return ts.createMethod(
       undefined,
-      [ts.createModifier(ts.SyntaxKind.PublicKeyword)],
+      [ts.createModifier(ts.SyntaxKind.PrivateKeyword)],
       undefined,
       'constructor',
       undefined,
       undefined,
-      [
-        ts.createParameter(
-          undefined,
-          undefined,
-          undefined,
-          'value',
-          undefined,
-          ts.createTypeReferenceNode('ShapeOf' + oautil.typenamify(key), [])
-        ),
-        ts.createParameter(
-          undefined,
-          undefined,
-          undefined,
-          'opts',
-          ts.createToken(ts.SyntaxKind.QuestionToken),
-          ts.createTypeReferenceNode(fromLib('make', 'MakeOptions'), [])
-        )
-      ],
+      [],
       undefined,
       ts.createBlock([
-        ts.createExpressionStatement(ts.createCall(ts.createIdentifier('super'), [], [])),
-        ts.createExpressionStatement(
-          ts.createCall(
-            ts.createPropertyAccess(ts.createIdentifier('Object'), 'assign'),
-            [],
-            [
-              ts.createIdentifier('this'),
-              ts.createCall(
-                ts.createPropertyAccess(
-                  ts.createCall(ts.createIdentifier('build' + oautil.typenamify(key)), undefined, [
-                    ts.createIdentifier('value'),
-                    ts.createIdentifier('opts')
-                  ]),
-                  ts.createIdentifier('success')
-                ),
-                undefined,
-                []
-              )
-            ]
-          )
-        )
+        ts.createExpressionStatement(ts.createCall(ts.createIdentifier('super'), [], []))
       ])
     );
   }
@@ -717,11 +680,66 @@ export function run(options: Options) {
         ts.createTypeReferenceNode(oautil.typenamify(key), [])
       ]),
       ts.createBlock([
+        ts.createVariableStatement(
+          [],
+          ts.createVariableDeclarationList(
+            [
+              ts.createVariableDeclaration(
+                'make',
+                undefined,
+                ts.createCall(ts.createIdentifier('build' + oautil.typenamify(key)), undefined, [
+                  ts.createIdentifier('value'),
+                  ts.createIdentifier('opts')
+                ])
+              )
+            ],
+            ts.NodeFlags.Const
+          )
+        ),
         ts.createReturn(
-          ts.createCall(ts.createIdentifier('make' + oautil.typenamify(key)), undefined, [
-            ts.createIdentifier('value'),
-            ts.createIdentifier('opts')
-          ])
+          ts.createConditional(
+            ts.createCall(
+              ts.createPropertyAccess(ts.createIdentifier('make'), ts.createIdentifier('isError')),
+              [],
+              []
+            ),
+            ts.createCall(
+              ts.createPropertyAccess(
+                ts.createPropertyAccess(ts.createPropertyAccess(runtimeLibrary, 'make'), 'Make'),
+                'error'
+              ),
+              undefined,
+              [ts.createPropertyAccess(ts.createIdentifier('make'), 'errors')]
+            ),
+            ts.createCall(
+              ts.createPropertyAccess(
+                ts.createPropertyAccess(ts.createPropertyAccess(runtimeLibrary, 'make'), 'Make'),
+                'ok'
+              ),
+              undefined,
+              [
+                ts.createCall(
+                  ts.createPropertyAccess(ts.createIdentifier('Object'), 'assign'),
+                  [],
+                  [
+                    ts.createCall(
+                      ts.createIdentifier('new ' + oautil.typenamify(key)),
+                      undefined,
+                      []
+                    ),
+                    ts.createCall(
+                      ts.createPropertyAccess(
+                        ts.createIdentifier('make'),
+                        ts.createIdentifier('success')
+                      ),
+                      undefined,
+                      []
+                    )
+                  ]
+                )
+              ]
+            )
+          )
         )
       ])
     );
@@ -729,7 +747,7 @@ export function run(options: Options) {
 
   function generateClassBuiltindMembers(key: string) {
     return [
-      generateClassConstructor(key),
+      generateClassConstructor(),
       generateReflectionProperty(key),
       generateClassMakeMethod(key)
     ];
@@ -1262,8 +1280,7 @@ export function run(options: Options) {
     );
   }
 
-  function generateTopLevelClassMaker(key: string, constructor: string) {
-    const makerFun = 'createMakerWith';
+  function generateTopLevelClassMaker(key: string) {
     const shape = 'ShapeOf' + oautil.typenamify(key);
     return ts.createVariableStatement(
       [ts.createModifier(ts.SyntaxKind.ExportKeyword)],
@@ -1275,7 +1292,7 @@ export function run(options: Options) {
               ts.createTypeReferenceNode(shape, []),
               ts.createTypeReferenceNode(oautil.typenamify(key), [])
             ]),
-            makeCall(makerFun, [ts.createIdentifier(constructor)])
+            ts.createPropertyAccess(ts.createIdentifier(oautil.typenamify(key)), 'make')
           )
         ],
         ts.NodeFlags.Const
@@ -1360,7 +1377,7 @@ export function run(options: Options) {
       generateBrand(key),
       generateValueClass(key, schema),
       generateTopLevelClassBuilder(key, schema),
-      generateTopLevelClassMaker(key, oautil.typenamify(key)),
+      generateTopLevelClassMaker(key),
       generateNamedTypeDefinitionAssignment(key, schema, generateIsA(oautil.typenamify(key)))
     ];
   }